+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### CHANGED
+- Moved props into a pointer group
+
+### DEPRECATED
+- accessing props directly from the scene
+
 ## [0.1.0] - 2025-04-13
 
 ### ADDED